fix: apply changes from code review
