+++ packages/website/src/components/lib/jsonSchema.ts
@@ -1,41 +1,9 @@
 import type { JSONSchema4 } from '@typescript-eslint/utils/json-schema';
 import type * as ts from 'typescript';
 
-import { getTypescriptOptions } from '../config/utils';
 import type { WebLinter } from '../linter/WebLinter';
 
-export function createCompilerOptions(
-  tsConfig: Record<string, unknown> = {},
-): ts.CompilerOptions {
-  const config = window.ts.convertCompilerOptionsFromJson(
-    {
-      // ts and monaco has different type as monaco types are not changing base on ts version
-      target: 'esnext',
-      module: 'esnext',
-      jsx: 'preserve',
-      ...tsConfig,
-      allowJs: true,
-      lib: Array.isArray(tsConfig.lib) ? tsConfig.lib : undefined,
-      moduleResolution: undefined,
-      plugins: undefined,
-      typeRoots: undefined,
-      paths: undefined,
-      moduleDetection: undefined,
-      baseUrl: undefined,
-    },
-    '/tsconfig.json',
-  );
-
-  const options = config.options;
-
-  if (!options.lib) {
-    options.lib = [window.ts.getDefaultLibFileName(options)];
-  }
-
-  return options;
-}
-
-export function getEslintSchema(linter: WebLinter): JSONSchema4 {
+export function getEslintJsonSchema(linter: WebLinter): JSONSchema4 {
   const properties: Record<string, JSONSchema4> = {};
 
   for (const [, item] of linter.rules) {
@@ -83,7 +51,38 @@ export function getEslintSchema(linter: WebLinter): JSONSchema4 {
   };
 }
 
-export function getTsConfigSchema(): JSONSchema4 {
+export function getTypescriptOptions(): ts.OptionDeclarations[] {
+  const allowedCategories = [
+    'Command-line Options',
+    'Projects',
+    'Compiler Diagnostics',
+    'Editor Support',
+    'Output Formatting',
+    'Watch and Build Modes',
+    'Source Map Options',
+  ];
+
+  const filteredNames = [
+    'moduleResolution',
+    'moduleDetection',
+    'plugins',
+    'typeRoots',
+    'jsx',
+  ];
+
+  return window.ts.optionDeclarations.filter(
+    item =>
+      (item.type === 'boolean' ||
+        item.type === 'list' ||
+        item.type instanceof Map) &&
+      item.description &&
+      item.category &&
+      !allowedCategories.includes(item.category.message) &&
+      !filteredNames.includes(item.name),
+  );
+}
+
+export function getTypescriptJsonSchema(): JSONSchema4 {
   const properties = getTypescriptOptions().reduce((options, item) => {
     if (item.type === 'boolean') {
       options[item.name] = {
